Admin side is developed

+++ src/pages/api/products/[id].js
@@ -2,44 +2,47 @@ import dbConnect from "../../../../lib/mongodb"
 import Product from "../../../../model/Product"
 
 export default async function handler(req, res) {
+    const {
+        method,
+        query: { id },
+        cookies
+    } = req
+    const token = cookies.token
+
     dbConnect()
 
-    const { method, query } = req
-    const { id } = query
-
-    // console.log(id)
-    // res.status(200).json({ message: "hello", id, method })
-    // return
-
-    // Delete Product.
-    if (method === "DELETE") {
+    if (method === "GET") {
         try {
-            await Product.findByIdAndDelete(id)
-            res.status(200).json("The product has been deleted!")
-        } catch (error) {
-            res.status(500).json(error)
+            const product = await Product.findById(id)
+            res.status(200).json(product)
+        } catch (err) {
+            res.status(500).json(err)
         }
     }
 
-    // Update Product.
     if (method === "PUT") {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json("Not authenticated!")
+        }
         try {
             const product = await Product.findByIdAndUpdate(id, req.body, {
                 new: true
             })
             res.status(200).json(product)
-        } catch (error) {
-            res.status(500).json(error)
+        } catch (err) {
+            res.status(500).json(err)
         }
     }
 
-    // Get one product.
-    if (method === "GET") {
+    if (method === "DELETE") {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json("Not authenticated!")
+        }
         try {
-            const product = await Product.findById(id)
-            res.status(200).json(product)
-        } catch (error) {
-            // res.status(500).json(error)
+            await Product.findByIdAndDelete(id)
+            res.status(200).json("The product has been deleted!")
+        } catch (err) {
+            res.status(500).json(err)
         }
     }
 }
